Move WarningsConfigurationBuilder to Infrastructure
Issue #5633
