+++ src/Microsoft.EntityFrameworkCore/Infrastructure/WarningsConfigurationBuilder.cs
@@ -3,11 +3,10 @@
 
 using System.Linq;
 using JetBrains.Annotations;
-using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Internal;
 using Microsoft.EntityFrameworkCore.Utilities;
 
-namespace Microsoft.EntityFrameworkCore
+namespace Microsoft.EntityFrameworkCore.Infrastructure
 {
     public class WarningsConfigurationBuilder
     {
